Remove bloat classes that are no longer needed
